requirements.txt
opencv-python
pillow
pytesseract
numpy
pandas
pdf2image
